backend/requirements.txt
Django==5.2.11
djangorestframework==3.16.1
django-cors-headers==4.9.0
python-dotenv==1.1.1
Pillow==11.3.0
PyMySQL==1.1.1
